pyproject.toml
[tool.poetry]
name = "hexlet-code"
version = "0.1.0"
description = ""
authors = ["Ivan <asgefes1@gmail.com>"]
readme = "README.md"
packages = [
 {include = "gendiff"},
]

[tool.poetry.dependencies]
python = "^3.12"

[tool.poetry.scripts]
gendiff = "gendiff.scripts.gendiff:main"


[tool.poetry.group.dev.dependencies]
flake8 = "^6.1.0"
pytest = "^7.4.3"
pytest-cov = "^4.1.0"
pyyaml = "^6.0.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
